refactor: move to types
