+++ src/index.ts
@@ -1,7 +1,8 @@
 import { createFetch } from "./base";
 
 export * from "./base";
-export type { ResponseType, ResponseMap, MappedType } from "./utils";
+
+export type { ResponseType, ResponseMap, MappedType } from "./types";
 
 // ref: https://github.com/tc39/proposal-global
 const _globalThis = (function () {
